fix: update sync and functional tests
